Split up miq_capacity into threee separate controllers
