+++ app/helpers/application_helper/title.rb
@@ -64,8 +64,12 @@ module ApplicationHelper
                 _(": Automation")
               when /^miq_policy/
                 _(": Control")
-              when /^miq_capacity/
-                _(": Optimize")
+              when /^miq_capacity_utilization/
+                _(": Utilization")
+              when /^miq_capacity_planning/
+                _(": Planning")
+              when /^miq_capacity_bottlenecks/
+                _(": Bottlenecks")
               when /^miq_request/
                 _(": Requests")
               when "manageiq/providers/ansible_tower/automation_manager/playbook"
