+++ app/helpers/application_helper/toolbar/miq_capacity_view.rb
@@ -12,7 +12,7 @@ class ApplicationHelper::Toolbar::MiqCapacityView < ApplicationHelper::Toolbar::
           'fa fa-file-text-o fa-lg',
           N_('Download this report in text format'),
           N_('Download as Text'),
-          :url       => "/\#{x_active_tree == :utilization_tree ? \"util_report\" : \"planning_report\"}_download",
+          :url       => "/report_download",
           :url_parms => "?typ=txt",
           :klass     => ApplicationHelper::Button::MiqCapacity),
         button(
@@ -20,7 +20,7 @@ class ApplicationHelper::Toolbar::MiqCapacityView < ApplicationHelper::Toolbar::
           'fa fa-file-text-o fa-lg',
           N_('Download this report in CSV format'),
           N_('Download as CSV'),
-          :url       => "/\#{x_active_tree == :utilization_tree ? \"util_report\" : \"planning_report\"}_download",
+          :url       => "/report_download",
           :url_parms => "?typ=csv",
           :klass     => ApplicationHelper::Button::MiqCapacity),
         button(
@@ -29,7 +29,7 @@ class ApplicationHelper::Toolbar::MiqCapacityView < ApplicationHelper::Toolbar::
           N_('Download this report in PDF format'),
           N_('Download as PDF'),
           :klass     => ApplicationHelper::Button::Pdf,
-          :url       => "/\#{x_active_tree == :utilization_tree ? \"util_report\" : \"planning_report\"}_download",
+          :url       => "/report_download",
           :url_parms => "?typ=pdf"),
       ]
     ),
