+++ app/presenters/tree_builder_utilization.rb
@@ -17,7 +17,7 @@ class TreeBuilderUtilization < TreeBuilderRegion
 
   def set_locals_for_render
     locals = super
-    locals.merge!(:onclick => "miqOnClickUtilization", :select_node => @selected_node.to_s)
+    locals.merge!(:select_node => @selected_node.to_s)
   end
 
   def root_options
