+++ app/views/bottlenecks/_bottlenecks_options.html.haml
@@ -1,4 +1,4 @@
-- url = url_for_only_path :action => 'bottleneck_tl_chooser'
+- url = url_for_only_path :action => 'tl_chooser'
 #tl_options_div
   %h3
     = _('Options')
