+++ app/views/bottlenecks/_bottlenecks_report.html.haml
