+++ app/views/bottlenecks/_bottlenecks_summary.html.haml
