+++ app/views/bottlenecks/_bottlenecks_tabs.html.haml
@@ -11,4 +11,4 @@
       = render :partial => "bottlenecks_report"
 
 :javascript
-  miq_tabs_init('bottlenecks_tabs', '/miq_capacity/change_tab/');
+  miq_tabs_init('bottlenecks_tabs', '/bottlenecks/change_tab/');
