+++ app/views/bottlenecks/_bottlenecks_tl_detail.html.haml
