+++ app/views/bottlenecks/index.html.haml
