--- app/views/miq_capacity/_bottlenecks_tree.html.haml
@@ -1,2 +1,0 @@
--# Div to hold the actual tree
-= render(:partial => 'shared/tree', :locals => {:tree => @bottlenecks_tree, :name => @bottlenecks_tree.name})
