--- app/views/miq_capacity/_utilization_tree.html.haml
@@ -1,2 +1,0 @@
--# Div to hold the actual tree
-= render(:partial => 'shared/tree', :locals => {:tree => @utilization_tree, :name => @utilization_tree.name})
