+++ app/views/planning/_planning_instructions.html.haml
