+++ app/views/planning/_planning_options.html.haml
@@ -1,4 +1,4 @@
-- url = url_for_only_path(:action => 'planning_option_changed')
+- url = url_for_only_path(:action => 'option_changed')
 -# Div to hold the planning options
 #planning_options_div
   = render :partial => "layouts/flash_msg"
