+++ app/views/planning/_planning_report.html.haml
