+++ app/views/planning/_planning_summary.html.haml
@@ -1,4 +1,4 @@
-- url = url_for_only_path(:action => 'planning_option_changed')
+- url = url_for_only_path(:action => 'option_changed')
 #planning_summary_div
   = render :partial => "layouts/flash_msg"
   - if @perf_record
