+++ app/views/planning/_planning_vm_profile.html.haml
