+++ app/views/utilization/_utilization_details.html.haml
