+++ app/views/utilization/_utilization_report.html.haml
