+++ app/views/utilization/_utilization_summary.html.haml
