+++ app/views/utilization/_utilization_tabs.html.haml
@@ -15,4 +15,4 @@
       = render :partial => "utilization_report"
 
 :javascript
-  miq_tabs_init('#utilization_tabs', '/miq_capacity/change_tab');
+  miq_tabs_init('#utilization_tabs', '/utilization/change_tab');
